requirements/base.txt
Django==5.0.1
django-cors-headers==4.3.1
django-health-check==3.17.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
